refactor: move pastIterations normalize and denormalize functions from models to reducers
