+++ app/assets/javascripts/models/beta/pastIteration.js
@@ -16,41 +16,3 @@ const serialize = (data) =>
     recursive: true,
     arrayRecursive: true,
   });
-
-export const normalizePastIterations = (pastIterations) => {
-  return pastIterations.reduce(
-    (acc, pastIteration) => {
-      const pastIterationId = pastIteration.iterationNumber;
-
-      acc.pastIterations.byId[pastIterationId] = { ...pastIteration };
-      acc.pastIterations.allIds.push(pastIterationId);
-
-      return acc;
-    },
-    {
-      pastIterations: {
-        byId: {},
-        allIds: [],
-      },
-    }
-  );
-};
-
-export const denormalizePastIterations = (pastIterations) => {
-  const normalizedPastIterations = pastIterations?.pastIterations;
-
-  if (
-    !normalizedPastIterations ||
-    normalizedPastIterations.allIds.length === 0
-  ) {
-    return [];
-  }
-
-  const denormalizedPastIterations = normalizedPastIterations.allIds.map(
-    (iterationId) => {
-      return normalizedPastIterations.byId[iterationId];
-    }
-  );
-
-  return denormalizedPastIterations;
-};
