--- app/assets/javascripts/selectors/pastIterations.js
@@ -1,3 +1,0 @@
-import { denormalizePastIterations } from "../models/beta/pastIteration";
-
-export const getIterations = (state) => denormalizePastIterations(state);
